1. Modified util_pedal_pico  a. Added inline function to count a number in a threshold gate.
2. Modified pedal_pico_buffer, pedal_pico_phaser, pedal_pico_sideband, pedal_pico_tremolo
 a. Fitted with Modification of util_pedal_pico
 b. Variable about Types and Casting

3. Modified pedal_pico_sustain
 a. Variable about Types and Casting
 b. Threshold Gate to Turn On or Off
 c. Switches to Change Output Gain
 d. Output Gain

+++ include/pedal_pico/pedal_pico_buffer.h
@@ -56,7 +56,7 @@ volatile uint8_t pedal_pico_buffer_delay_amplitude_interpolation_accum;
 volatile uint16_t pedal_pico_buffer_delay_time;
 volatile uint16_t pedal_pico_buffer_delay_time_interpolation;
 volatile uint16_t pedal_pico_buffer_delay_index;
-volatile int8_t pedal_pico_buffer_noise_gate_threshold;
+volatile uint16_t pedal_pico_buffer_noise_gate_threshold;
 volatile uint16_t pedal_pico_buffer_noise_gate_count;
 
 void pedal_pico_buffer_set();
