+++ include/pedal_pico/pedal_pico_phaser.h
@@ -53,7 +53,7 @@ volatile uint16_t pedal_pico_phaser_delay_index;
 volatile int32_t pedal_pico_phaser_osc_triangle_1_index; // May Have Negative Value Depending on osc_speed
 volatile uint16_t pedal_pico_phaser_osc_speed;
 volatile bool pedal_pico_phaser_osc_is_negative;
-volatile int8_t pedal_pico_phaser_osc_start_threshold;
+volatile uint16_t pedal_pico_phaser_osc_start_threshold;
 volatile uint16_t pedal_pico_phaser_osc_start_count;
 
 void pedal_pico_phaser_set();
